+++ integration/upstream_client_test.go
@@ -269,3 +269,21 @@ func createSnapshotCache(updates int, logger log.Logger) (resourcev2.TestSnapsho
 		NumClusters:  updates,
 	}, gcpcachev2.NewSnapshotCache(false, gcpcachev2.IDHash{}, gcpLogger{logger: logger})
 }
+
+type gcpLogger struct{}
+
+func (logger gcpLogger) Debugf(format string, args ...interface{}) {
+	log.Printf(format+"\n", args...)
+}
+
+func (logger gcpLogger) Infof(format string, args ...interface{}) {
+	log.Printf(format+"\n", args...)
+}
+
+func (logger gcpLogger) Warnf(format string, args ...interface{}) {
+	log.Printf(format+"\n", args...)
+}
+
+func (logger gcpLogger) Errorf(format string, args ...interface{}) {
+	log.Printf(format+"\n", args...)
+}
